Generalize 'module'
Goal: Module class a generic class describing either verilog
modules or VHDL entities.

Transfer verilog_module to sv/verilog_module, and make 'module' an alias
of it.
